pyproject.toml
[project]
name = "tab-vds-demo"
version = "0.1.0"
description = "Streamlit app for Tableau Web Authoring and VizQL Data Service integration"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "pandas>=2.3.2",
    "pyjwt>=2.10.1",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
]
